code separation for unauthenticated users

+++ orders/views.py
@@ -4,6 +4,7 @@ from django.contrib import messages
 from .models import Cart,CartItem,PlacedOrder
 from django.shortcuts import redirect, render, resolve_url
 from menus.models import MenuItem
+import json
 # Create your views here.
 
 #helper function for session retreival
@@ -24,60 +25,66 @@ def add_item(item):
     item.save()
     
 
-def remove_item(item):
+def decrease_item(item):
     item.quantity -=1
     item.total_price -= item.menu_item.price
     item.save()
   
 
 
+# add to cart in database for logged in users 
+# and just make a dict for unauthenticated users 
 
 def add_cart(request,menu_id):
+    # add cart for logged in user
      #check if cart exists or not (id not then create)
-    session_id = get_session(request)
     if request.user.is_authenticated:
         cust = Customers.objects.get(user= request.user)
         cart,created = Cart.objects.get_or_create(user=cust,defaults={
-            'session_id':session_id,
             'user':cust
         })
-    else:
-        cart,created = Cart.objects.get_or_create(session_id=session_id,defaults={
-            'session_id':session_id
-        })
+        menuitem = MenuItem.objects.get(id=menu_id)
+        menuitem.is_added =True
+        menuitem.save()
         
-
-    menuitem = MenuItem.objects.get(id=menu_id)
-    menuitem.is_added =True
-    menuitem.save()
-    item,item_created = CartItem.objects.get_or_create(menu_item=menuitem,cart=cart,defaults={
-        'menu_item':menuitem,
-        'quantity': 1,
-        'total_price':menuitem.price,
-        'cart':cart
-    })
-    cart.count+=1
-    cart.total = float(cart.total) + float(item.quantity * item.menu_item.price)
-    cart.save()
-    if not item_created:
-        add_item(item)
+        item,item_created = CartItem.objects.get_or_create(menu_item=menuitem,cart=cart,defaults={
+            'menu_item':menuitem,'quantity': 1,'total_price':menuitem.price,'cart':cart})
+        cart.count+=1
+        cart.total = float(cart.total) + float(item.quantity * item.menu_item.price)
+        cart.save()
+        print(item)
+        if not item_created:
+            add_item(item)
     
+    else:
+        # get data from dictionary
+        data = request.COOKIES.get('cart')
+        print(data)
+        cart = {
+            'user':'',
+            'count':'',
+            'total':'',
+            'updated':'',
+            'timestamp':''
+            }
 
     return redirect('/view_cart')
 
 
 def remove_cart(request,menu_id):
-    session_id= get_session(request)
-    cart = Cart.objects.get(session_id=session_id)
-    item = CartItem.objects.get(menu_item__id = menu_id,cart=cart)
-    cart.count-=1
-    cart.total = float(cart.total) - float(item.quantity * item.menu_item.price)
-    cart.save()
-    if item.quantity == 0:
-        item.delete()
-    else:
-        remove_item(item)
-    
+    if request.user.is_authenticated:
+        cust=Customers.objects.get(user=request.user)
+        cart = Cart.objects.get(user=cust)
+        item = CartItem.objects.get(menu_item__id = menu_id,cart=cart)
+        print(item)
+        cart.count-=1
+        cart.total = float(cart.total) - float(item.quantity * item.menu_item.price)
+        cart.save()
+        if item.quantity == 0:
+            item.delete()
+        else:
+            decrease_item(item)
+        
     return redirect('/view_cart')
 
 
@@ -87,13 +94,16 @@ def view_cart(request):
             if request.user.is_authenticated:
                 cust = Customers.objects.get(user= request.user)
                 cart=Cart.objects.get(user=cust)
-
+                cartitem=CartItem.objects.filter(cart=cart)
+                total_value= cart.total
             else:
-                cart=Cart.objects.get(session_id=get_session(request))
-                print('calling from else')
-            cartitem=CartItem.objects.filter(cart=cart)
-            print(cartitem)
-            total_value= cart.total
+                cartitem = {}
+                total_value = 0
+                data = request.COOKIES.get('cart')
+                print(data)
+                #get data from cookies as user is not logged in
+                # cart=Cart.objects.get(session_id=get_session(request))
+            
         except ObjectDoesNotExist:
             messages.add_message(request,messages.INFO,'Cart is empty!')
             total_value = 0
@@ -111,13 +121,17 @@ def view_cart(request):
 
 def remove_item(request,item_id):
     if request.method == 'GET':
-        cartitem=CartItem.objects.get(id=item_id)
-        cust = Customers.objects.get(user= request.user)
-        cart = Cart.objects.get(user=cust)
-        cart.total -= cartitem.total_price
-        cart.count -=cartitem.quantity
-        cart.save()
-        cartitem.delete()
+        if request.user.is_authenticated:
+            cartitem=CartItem.objects.get(id=item_id)
+            cust = Customers.objects.get(user= request.user)
+            cart = Cart.objects.get(user=cust)
+            cart.total -= cartitem.total_price
+            cart.count -=cartitem.quantity
+            cart.save()
+            cartitem.delete()
+        else:
+            #is user is not logged in
+            pass
         return redirect('/view_cart')
